@@ -11,6 +11,7 @@
     "make": "electron-forge make",
     "release": "electron-forge publish",
     "lint": "eslint . --ext js,ts",
+    "fix": "eslint --ext .ts,.tsx --fix .",
     "test": "jest",
     "rebuild": "./node_modules/.bin/electron-rebuild"
   },
@@ -53,8 +54,8 @@
     "@types/react-dom": "17.0.8",
     "@types/react-syntax-highlighter": "^15.5.6",
     "@types/styled-components": "5.1.10",
-    "@typescript-eslint/eslint-plugin": "4.27.0",
-    "@typescript-eslint/parser": "4.27.0",
+    "@typescript-eslint/eslint-plugin": "^5.57.1",
+    "@typescript-eslint/parser": "^5.57.1",
     "@vercel/webpack-asset-relocator-loader": "^1.7.3",
     "@zeit/webpack-asset-relocator-loader": "^0.8.0",
     "ajv": "^7",
@@ -64,14 +65,17 @@
     "cross-env": "7.0.3",
     "electron": "^23.2.1",
     "electron-rebuild": "^3.2.9",
-    "eslint": "7.29.0",
+    "eslint": "^8.37.0",
+    "eslint-config-airbnb": "^19.0.4",
     "eslint-config-prettier": "8.3.0",
     "eslint-config-standard": "16.0.3",
-    "eslint-plugin-import": "2.23.4",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-node": "11.1.0",
     "eslint-plugin-prettier": "3.4.0",
     "eslint-plugin-promise": "5.1.0",
-    "eslint-plugin-react": "7.24.0",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-standard": "5.0.0",
     "file-loader": "^6.2.0",
     "html2plaintext": "^2.1.4",
