@@ -1,43 +1,44 @@
-import os from 'os'
-const applescript = require('applescript')
+import os from 'os';
+
+const applescript = require('applescript');
 
 function runAppleScript(script: string) {
   return new Promise((resolve, reject) => {
     applescript.execString(script, (err: any, rtn: any) => {
       if (err) {
-        reject(err)
-        return
+        reject(err);
+        return;
       }
-      resolve(rtn)
-    })
-  })
+      resolve(rtn);
+    });
+  });
 }
 
 // TODO: 这种方法在vscode上哪怕没有选中也会复制一行
-export function getSelection() {  
+export function getSelection() {
   const script = `
     tell application "System Events" to keystroke "c" using {command down}
     delay 0.1
     set selectedText to the clipboard
-  `
-  return runAppleScript(script)
+  `;
+  return runAppleScript(script);
 }
 
-export function applySelection() {  
+export function applySelection() {
   const script = `
     tell application "System Events" to keystroke "v" using {command down}
     delay 1
     set selectedText to the clipboard
-  `
-  return runAppleScript(script)
+  `;
+  return runAppleScript(script);
 }
 
 export function getRecentApp() {
   const script = `
   tell application "System Events"
     set frontmostAppName to displayed name of first application process whose frontmost is true        
-  end tell`
-  return runAppleScript(script)
+  end tell`;
+  return runAppleScript(script);
 }
 
 export function activeApp(app: string) {
@@ -45,8 +46,8 @@ export function activeApp(app: string) {
     tell application "${app}"
       activate
     end tell  
-    `
-  return runAppleScript(script)    
+    `;
+  return runAppleScript(script);
 }
 
 export function getBrowserContnet() {
@@ -59,6 +60,6 @@ export function getBrowserContnet() {
     end tell
     
   end tell
-  `
-  return runAppleScript(script)  
+  `;
+  return runAppleScript(script);
 }
