@@ -1,11 +1,11 @@
-import { ChatGPTAPI } from 'chatgpt'
+import { ChatGPTAPI } from 'chatgpt';
 
 export async function chat() {
   const api = new ChatGPTAPI({
-    apiKey: process.env.OPENAI_API_KEY || ""
-  })
+    apiKey: process.env.OPENAI_API_KEY || '',
+  });
 
   try {
-    await api.sendMessage('Hello World!')
-  } catch(e) {}
+    await api.sendMessage('Hello World!');
+  } catch (e) {}
 }
