@@ -1,12 +1,13 @@
-import { IpcMainInvokeEvent, ipcMain, BrowserWindow, clipboard } from 'electron'
-import { Singleton } from "./global"
-import { activeApp, applySelection } from './os'
-import { Logger } from './util'
+import {
+  IpcMainInvokeEvent, ipcMain, BrowserWindow, clipboard,
+} from 'electron';
+import { Singleton } from './global';
+import { activeApp, applySelection } from './os';
+import { Logger } from './util';
 
+const clipboardWatcher = require('electron-clipboard-watcher');
 
-const clipboardWatcher = require('electron-clipboard-watcher')
-
-export function listen(win:  BrowserWindow | null ) {
+export function listen(win: BrowserWindow | null) {
   clipboardWatcher({
     // (optional) delay in ms between polls
     watchDelay: 200,
@@ -15,30 +16,29 @@ export function listen(win:  BrowserWindow | null ) {
     // handler for when text data is copied into the clipboard
     onTextChange(text: string) {
       if (Singleton.getInstance().getCopyFromElectron()) {
-        Logger.log('来自chat内容拷贝')
-        Singleton.getInstance().setCopyStateSource(false)
-        return
+        Logger.log('来自chat内容拷贝');
+        Singleton.getInstance().setCopyStateSource(false);
       }
       // TODO: 粘贴板内容变化弹窗，感觉没必要保留，会对用户体验造成非常大的影响
       // setWindowVisile(true)
       //   win?.webContents.send('clipboard_change', text)
     },
-  })
+  });
 
   ipcMain.handle(
     'attemptChange',
-    async function (
+    async (
       event: IpcMainInvokeEvent,
-      changes: string
-    ) {
+      changes: string,
+    ) => {
       try {
-        clipboard.writeText(changes)
-        await activeApp(Singleton.getInstance().getRecentApp())
-        const result = await applySelection()
-        Logger.log('应用改变:', result)
-      } catch(e) {
-        Logger.error(e)
+        clipboard.writeText(changes);
+        await activeApp(Singleton.getInstance().getRecentApp());
+        const result = await applySelection();
+        Logger.log('应用改变:', result);
+      } catch (e) {
+        Logger.error(e);
       }
-    }
-  )  
+    },
+  );
 }
