@@ -1,29 +1,30 @@
 export class Singleton {
-  private static instance: Singleton
-  private static copyFromElectron: boolean = false
-  private static recentApp: string
+  private static instance: Singleton;
+
+  private static copyFromElectron: boolean = false;
+
+  private static recentApp: string;
 
   public static getInstance(): Singleton {
     if (!Singleton.instance) {
-      Singleton.instance = new Singleton()
+      Singleton.instance = new Singleton();
     }
-    return Singleton.instance
+    return Singleton.instance;
   }
 
   public setRecentApp(app: string) {
-    Singleton.recentApp = app
+    Singleton.recentApp = app;
   }
 
   public setCopyStateSource(fromElectron: boolean) {
-    Singleton.copyFromElectron = fromElectron
+    Singleton.copyFromElectron = fromElectron;
   }
 
   public getCopyFromElectron() {
-    return Singleton.copyFromElectron
+    return Singleton.copyFromElectron;
   }
 
   public getRecentApp() {
-    return Singleton.recentApp
+    return Singleton.recentApp;
   }
-
 }
