@@ -1,21 +1,22 @@
-import { 
-  app, 
-  BrowserWindow, 
+import {
+  app,
+  BrowserWindow,
   ipcMain,
-} from 'electron'
-import { setupStoreHandlers } from './store'
-import { Logger } from './util'
-import { setWindowVisile } from './window'
-import { listen as setupShortcutHandlers } from './shortcuts'
-import { listen as setupClipboardHandlers } from './clipboard'
-require('./server')
+} from 'electron';
+import { setupStoreHandlers } from './store';
+import { Logger } from './util';
+import { setWindowVisile } from './window';
+import { listen as setupShortcutHandlers } from './shortcuts';
+import { listen as setupClipboardHandlers } from './clipboard';
 
-declare const MAIN_WINDOW_WEBPACK_ENTRY: string
-declare const MAIN_WINDOW_PRELOAD_WEBPACK_ENTRY: string
-let win: BrowserWindow | null
+require('./server');
 
-function createWindow () {
-  win = new BrowserWindow({    
+declare const MAIN_WINDOW_WEBPACK_ENTRY: string;
+declare const MAIN_WINDOW_PRELOAD_WEBPACK_ENTRY: string;
+let win: BrowserWindow | null;
+
+function createWindow() {
+  win = new BrowserWindow({
     // useContentSize: true,
     resizable: false,
     width: 800,
@@ -26,59 +27,59 @@ function createWindow () {
     backgroundColor: '#00000000',
     skipTaskbar: true,
     webPreferences: {
-      webSecurity: false,      
+      webSecurity: false,
       backgroundThrottling: false,
       contextIsolation: true,
       webviewTag: true,
       nodeIntegration: true,
-      preload: MAIN_WINDOW_PRELOAD_WEBPACK_ENTRY
+      preload: MAIN_WINDOW_PRELOAD_WEBPACK_ENTRY,
       // enableRemoteModule: true,
       // preload: path.join(__static, "preload.js")
-    }
-  })
-  if(process.env.NODE_ENV !== 'production') {
-    win?.webContents.openDevTools({
-      mode:'bottom'
-    })
-  } 
-  win.loadURL(MAIN_WINDOW_WEBPACK_ENTRY)
-  win.on('closed', () => {
-    win = null
-  })
-  win.on("blur", () => {
-    setWindowVisile({
-      win, 
-      visible: false
-    })
+    },
   });
-  registerListeners()  
+  if (process.env.NODE_ENV !== 'production') {
+    win?.webContents.openDevTools({
+      mode: 'bottom',
+    });
+  }
+  win.loadURL(MAIN_WINDOW_WEBPACK_ENTRY);
+  win.on('closed', () => {
+    win = null;
+  });
+  win.on('blur', () => {
+    setWindowVisile({
+      win,
+      visible: false,
+    });
+  });
+  registerListeners();
 }
 
-async function registerListeners () {  
+async function registerListeners() {
   ipcMain.on('message', (_, message) => {
-    Logger.log(message)    
-  })
+    Logger.log(message);
+  });
   ipcMain.on('win_ignore_mouse', (_, ignore) => {
-    win?.setIgnoreMouseEvents(ignore, { forward: true })
-  })  
-  setupClipboardHandlers(win)
-  setupShortcutHandlers(win)
-  setupStoreHandlers()
+    win?.setIgnoreMouseEvents(ignore, { forward: true });
+  });
+  setupClipboardHandlers(win);
+  setupShortcutHandlers(win);
+  setupStoreHandlers();
 }
 
 app.on('ready', createWindow)
   .whenReady()
   .then()
-  .catch(e => console.error(e))
+  .catch((e) => console.error(e));
 
 app.on('window-all-closed', () => {
   if (process.platform !== 'darwin') {
-    app.quit()
+    app.quit();
   }
-})
+});
 
 app.on('activate', () => {
   if (BrowserWindow.getAllWindows().length === 0) {
-    createWindow()
+    createWindow();
   }
-})
+});
