@@ -4,45 +4,45 @@ import {
   ipcMain,
   clipboard,
   globalShortcut,
-} from 'electron'
+} from 'electron';
 
-import { getSelection, getRecentApp } from './os'
-import { Logger } from './util'
-import { setWindowVisile } from './window'
-import { Singleton } from "./global"
+import { getSelection, getRecentApp } from './os';
+import { Logger } from './util';
+import { setWindowVisile } from './window';
+import { Singleton } from './global';
 
 export function listen(win: BrowserWindow | null) {
   globalShortcut.register('CommandOrControl+k', async () => {
-    const visible = win?.isVisible() && win.isFocused()
+    const visible = win?.isVisible() && win.isFocused();
     if (!visible) {
       try {
-        const app = (await getRecentApp()) as string
-        Logger.log('appName', app)
-        Singleton.getInstance().setRecentApp(app)
+        const app = (await getRecentApp()) as string;
+        Logger.log('appName', app);
+        Singleton.getInstance().setRecentApp(app);
         // 每次获取完后需要清空一下剪切板，否则判断会有问题
-        const selection = await getSelection()
-        Logger.log('selectionTxt', selection)        
+        const selection = await getSelection();
+        Logger.log('selectionTxt', selection);
         win?.webContents.send('selection_change', {
           txt: selection,
-          app
-        })
+          app,
+        });
         setWindowVisile({
           win,
           visible: true,
-        })
-        clipboard.writeText('')
+        });
+        clipboard.writeText('');
       } catch (e) {
-        Logger.error(e)
+        Logger.error(e);
         setWindowVisile({
           win,
           visible: true,
-        })
+        });
       }
-      return
+      return;
     }
     setWindowVisile({
       win,
       visible: false,
-    })
-  })
+    });
+  });
 }
