@@ -1,16 +1,16 @@
-import { IpcMainInvokeEvent, ipcMain } from 'electron'
-import Store from 'electron-store'
+import { IpcMainInvokeEvent, ipcMain } from 'electron';
+import Store from 'electron-store';
 
-const store = new Store()
+const store = new Store();
 
 export function setupStoreHandlers() {
   ipcMain.handle(
     'setStore',
-    async function (
+    async (
       event: IpcMainInvokeEvent,
-      { key, blob }: { key: string; blob: any }
-    ) {
-      store.set(key, blob)
-    }
-  )
+      { key, blob }: { key: string; blob: any },
+    ) => {
+      store.set(key, blob);
+    },
+  );
 }
