@@ -1,16 +1,16 @@
-import { BrowserWindow } from 'electron'
+import { BrowserWindow } from 'electron';
 
 export const setWindowVisile = (opt: { win: BrowserWindow | null, visible?: boolean }) => {
-  const { win, visible } = opt
+  const { win, visible } = opt;
   if (!visible) {
-    win?.hide()
-    win?.blur()
-    return
+    win?.hide();
+    win?.blur();
+    return;
   }
   // win?.setAlwaysOnTop(true)
   win?.setVisibleOnAllWorkspaces(true, {
     visibleOnFullScreen: true,
-  })
-  win?.focus()
-  win?.show()
-}
+  });
+  win?.focus();
+  win?.show();
+};
