@@ -1,12 +1,14 @@
-import { logo, chat, translate, code, post } from './images'
+import {
+  logo, chat, translate, code, post,
+} from './images';
 
-export const Casual = 'Casual'
-export const Translate = 'Translate'
-export const Summarize = 'Summarize'
-export const Prettier = 'Prettier'
-export const Analyze = 'Analyze'
+export const Casual = 'Casual';
+export const Translate = 'Translate';
+export const Summarize = 'Summarize';
+export const Prettier = 'Prettier';
+export const Analyze = 'Analyze';
 
-export const BuiltInPlugins = [  
+export const BuiltInPlugins = [
   {
     logo: chat,
     id: Casual,
@@ -34,4 +36,4 @@ export const BuiltInPlugins = [
     loading: false,
     inputDisable: true,
   },
-]
+];
