@@ -1,12 +1,12 @@
 // https://imgloc.com/
-export const brand = 'https://i.imgur.com/T5ELmVC.png'
-export const logo = 'https://i.postimg.cc/tTJ3yHM9/pointer.png'
-export const robot = 'https://www.1ptai.com/logo_v2.png'
+export const brand = 'https://i.imgur.com/T5ELmVC.png';
+export const logo = 'https://i.postimg.cc/tTJ3yHM9/pointer.png';
+export const robot = 'https://www.1ptai.com/logo_v2.png';
 
-export const chat = 'https://i.328888.xyz/2023/04/05/i8Skpc.png'
-export const post = 'https://i.328888.xyz/2023/04/05/i8S8Tz.png'
-export const code = 'https://i.328888.xyz/2023/04/05/i8SjEq.png'
-export const translate = 'https://i.328888.xyz/2023/04/05/i8SNkw.png'
-export const loadingGif = 'https://superstorefinder.net/support/wp-content/uploads/2018/01/elastic.gif'
-export const searchLogo = 'https://i.328888.xyz/2023/04/06/iI8ySQ.png'
-export const logoLoading = 'http://superstorefinder.net/img/ripple-loader.svg'
+export const chat = 'https://i.328888.xyz/2023/04/05/i8Skpc.png';
+export const post = 'https://i.328888.xyz/2023/04/05/i8S8Tz.png';
+export const code = 'https://i.328888.xyz/2023/04/05/i8SjEq.png';
+export const translate = 'https://i.328888.xyz/2023/04/05/i8SNkw.png';
+export const loadingGif = 'https://superstorefinder.net/support/wp-content/uploads/2018/01/elastic.gif';
+export const searchLogo = 'https://i.328888.xyz/2023/04/06/iI8ySQ.png';
+export const logoLoading = 'http://superstorefinder.net/img/ripple-loader.svg';
