@@ -1,8 +1,8 @@
-import { configureStore, combineReducers } from '@reduxjs/toolkit'
-import chatReducer, { initialState as chatInitState } from '../features/chat/chatSlice'
-import presetReducer, { initialState as presetInitState } from '../features/preset/presetSlice'
-import settingReducer, { initialState as settingInitState } from '../features/setting/settingSlice'
-import clipboardReducer, { initialState as clipboardInitState } from '../features/clipboard/clipboardSlice'
+import { configureStore, combineReducers } from '@reduxjs/toolkit';
+import chatReducer, { initialState as chatInitState } from '../features/chat/chatSlice';
+import presetReducer, { initialState as presetInitState } from '../features/preset/presetSlice';
+import settingReducer, { initialState as settingInitState } from '../features/setting/settingSlice';
+import clipboardReducer, { initialState as clipboardInitState } from '../features/clipboard/clipboardSlice';
 
 export type StateType = {
   chat: typeof chatInitState;
@@ -15,19 +15,19 @@ export const initialState: StateType = {
   chat: chatInitState,
   preset: presetInitState,
   setting: settingInitState,
-  clipboard: clipboardInitState
-} 
+  clipboard: clipboardInitState,
+};
 
 const store = configureStore({
   reducer: combineReducers({
     chat: chatReducer,
     preset: presetReducer,
     setting: settingReducer,
-    clipboard: clipboardReducer
+    clipboard: clipboardReducer,
   }),
   preloadedState: initialState,
-})
+});
 
 export type RootState = ReturnType<typeof store.getState>
 export type AppDispatch = typeof store.dispatch
-export default store
+export default store;
