@@ -1,9 +1,9 @@
-import { render } from '@testing-library/react'
-import { Button } from './index'
+import { render } from '@testing-library/react';
+import { Button } from './index';
 
 test('button should renders', () => {
-  const { getByText } = render(<Button>ButtonContent</Button>)
+  const { getByText } = render(<Button>ButtonContent</Button>);
 
-  expect(getByText('ButtonContent')).toBeTruthy()
-  expect(getByText('ButtonContent')).toHaveAttribute('type', 'button')
-})
+  expect(getByText('ButtonContent')).toBeTruthy();
+  expect(getByText('ButtonContent')).toHaveAttribute('type', 'button');
+});
