@@ -1,11 +1,11 @@
-import { ReactNode, ButtonHTMLAttributes } from 'react'
+import { ReactNode, ButtonHTMLAttributes } from 'react';
 
-import { Container } from './styles'
+import { Container } from './styles';
 
 type ButtonProps = {
   children: ReactNode;
 } & ButtonHTMLAttributes<HTMLButtonElement>;
 
-export function Button (props: ButtonProps) {
-  return <Container type="button" {...props} />
+export function Button(props: ButtonProps) {
+  return <Container type="button" {...props} />;
 }
