@@ -1,56 +1,56 @@
-import { useEffect, useState } from 'react'
-import { Divider, Button } from 'antd'
-import { useAppSelector, useAppDispatch } from '../../app/hooks'
+import { useEffect, useState } from 'react';
+import { Divider, Button } from 'antd';
 
-import ReactMarkdown from 'react-markdown'
-import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter'
-import { atomDark } from 'react-syntax-highlighter/dist/esm/styles/prism'
+import ReactMarkdown from 'react-markdown';
+import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
+import { atomDark } from 'react-syntax-highlighter/dist/esm/styles/prism';
+import { useAppSelector, useAppDispatch } from '../../app/hooks';
 
-import { fetchChatResp } from '../../features/chat/chatSlice'
-import { presetMap, PresetType } from '../../features/preset/presetSlice'
-import { BuiltInPlugins } from '../../app/constants'
+import { fetchChatResp } from '../../features/chat/chatSlice';
+import { presetMap, PresetType } from '../../features/preset/presetSlice';
+import { BuiltInPlugins } from '../../app/constants';
 
 export function ChatPanel() {
-  const chatState = useAppSelector(state => state.chat)
-  const presetState = useAppSelector(state => state.preset)
-  const clipboardState = useAppSelector(state => state.clipboard)    
-  const [showSelection, setShowSelection] = useState<boolean>(false)
-  const [usePlugin, setUsePlugin] = useState<PresetType>()
-  const dispatch = useAppDispatch()
+  const chatState = useAppSelector((state) => state.chat);
+  const presetState = useAppSelector((state) => state.preset);
+  const clipboardState = useAppSelector((state) => state.clipboard);
+  const [showSelection, setShowSelection] = useState<boolean>(false);
+  const [usePlugin, setUsePlugin] = useState<PresetType>();
+  const dispatch = useAppDispatch();
 
   useEffect(() => {
     // TODO: should use id
-    const plugin = BuiltInPlugins.filter(item => presetState.currentPreset === item.title)[0]
-    setUsePlugin(plugin)
-  }, [presetState.currentPreset])
+    const plugin = BuiltInPlugins.filter((item) => presetState.currentPreset === item.title)[0];
+    setUsePlugin(plugin);
+  }, [presetState.currentPreset]);
 
   useEffect(() => {
-    setShowSelection(!!clipboardState.selectTxt && !!clipboardState.selectApp && !!usePlugin?.inputDisable)
-  },[clipboardState.selectTxt, clipboardState.selectApp, usePlugin?.inputDisable])
-  
+    setShowSelection(!!clipboardState.selectTxt && !!clipboardState.selectApp && !!usePlugin?.inputDisable);
+  }, [clipboardState.selectTxt, clipboardState.selectApp, usePlugin?.inputDisable]);
+
   const atemptChange = () => {
     // @ts-ignore
     // 删除markdown语法标识
-    window.Main.attemptChange(chatState.resp.replace(/^`{3}[^\n]+|`{3}$/g, ''))
-  }
+    window.Main.attemptChange(chatState.resp.replace(/^`{3}[^\n]+|`{3}$/g, ''));
+  };
 
   const doRequest = (txt: string) => {
     // @ts-ignore
-    const qaTpl = presetMap[presetState.currentPreset] as string
-    const qa = qaTpl + txt
-    setShowSelection(false)
+    const qaTpl = presetMap[presetState.currentPreset] as string;
+    const qa = qaTpl + txt;
+    setShowSelection(false);
     dispatch(
       fetchChatResp({
         question: qa,
-      })
-    )
-  }
+      }),
+    );
+  };
 
   const cancelRequest = () => {
-    setShowSelection(false)
-  }  
-  
-  return chatState.visible && (showSelection || chatState.resp)? (
+    setShowSelection(false);
+  };
+
+  return chatState.visible && (showSelection || chatState.resp) ? (
     <>
       <Divider style={{ margin: 0 }} />
       {showSelection ? (
@@ -60,10 +60,13 @@ export function ChatPanel() {
             {/* <span style={styles.selectTxt}>{`${
               String(clipboardState?.selectTxt)?.substring(0, 50) + '...'
             }`}</span>{' '} */}
-            in{' '}
+            in
+            {' '}
             <span
               style={styles.selectApp}
-            >{`${clipboardState.selectApp}`}</span>
+            >
+              {`${clipboardState.selectApp}`}
+            </span>
             ?
           </span>
           {/* 立即发问 */}
@@ -71,7 +74,7 @@ export function ChatPanel() {
             type="text"
             ghost
             danger
-            style={{ color: "rgb(255, 90, 0)" }}
+            style={{ color: 'rgb(255, 90, 0)' }}
             onClick={() => doRequest(clipboardState.selectTxt)}
           >
             Yes
@@ -87,11 +90,13 @@ export function ChatPanel() {
           <ReactMarkdown
             children={chatState.resp}
             components={{
-              code({ node, inline, className, children, ...props }) {
-                const match = /language-(\w+)/.exec(className || '')
+              code({
+                node, inline, className, children, ...props
+              }) {
+                const match = /language-(\w+)/.exec(className || '');
                 return !inline && match ? (
                   <SyntaxHighlighter
-                    showLineNumbers={true}
+                    showLineNumbers
                     children={String(children).replace(/\n$/, '')}
                     // @ts-ignore
                     style={atomDark}
@@ -103,14 +108,14 @@ export function ChatPanel() {
                   <code className={className} {...props}>
                     {children}
                   </code>
-                )
+                );
               },
             }}
           />
           {chatState.resp ? (
             <>
               <Divider style={{ margin: '25 0 0 0' }} />
-              <Button type="text" block onClick={atemptChange} style={{ fontSize: 12, fontWeight: "bold" }}>
+              <Button type="text" block onClick={atemptChange} style={{ fontSize: 12, fontWeight: 'bold' }}>
                 Attempt Change
               </Button>
             </>
@@ -118,10 +123,10 @@ export function ChatPanel() {
         </div>
       </div>
     </>
-  ) : null
+  ) : null;
 }
 
-const padding = 15
+const padding = 15;
 const styles = {
   selectionWrap: {
     backgroundColor: 'rgb(240 240 240)',
@@ -137,7 +142,7 @@ const styles = {
   selectApp: {
     fontStyle: 'italic',
     fontSize: 15,
-    fontWeight: "bold",
+    fontWeight: 'bold',
     // marginLeft: 4,
     // marginRight: 4,
     // color: 'rgb(255, 90, 0)',
@@ -160,4 +165,4 @@ const styles = {
     marginLeft: 30,
     overflow: 'auto',
   },
-}
+};
