@@ -1,6 +1,8 @@
-import { Avatar, List, Skeleton, Divider } from 'antd'
-import { useAppSelector, useAppDispatch } from '../../app/hooks'
-import { setListVisible as setPresetListVisible } from '../../features/preset/presetSlice'
+import {
+  Avatar, List, Skeleton, Divider,
+} from 'antd';
+import { useAppSelector, useAppDispatch } from '../../app/hooks';
+import { setListVisible as setPresetListVisible } from '../../features/preset/presetSlice';
 
 export interface PresetType {
   logo: string
@@ -13,11 +15,11 @@ interface Props {
   onPresetChange: (preset: string) => unknown
 }
 
-const padding = 15
+const padding = 15;
 export function Preset(props: Props) {
-  const { onPresetChange } = props
-  const presetState = useAppSelector((state) => state.preset)
-  const dispatch = useAppDispatch()
+  const { onPresetChange } = props;
+  const presetState = useAppSelector((state) => state.preset);
+  const dispatch = useAppDispatch();
 
   return presetState.listVisible ? (
     <>
@@ -30,7 +32,7 @@ export function Preset(props: Props) {
           itemLayout="horizontal"
           // loadMore={loadMore}
           dataSource={presetState.builtInPlugins}
-          renderItem={item => (
+          renderItem={(item) => (
             <List.Item
               className="ant-list-item"
               actions={[
@@ -38,9 +40,9 @@ export function Preset(props: Props) {
                   Edit
                 </a>,
               ]}
-              onClick={() => { 
-                onPresetChange(item.title) 
-                dispatch(setPresetListVisible(false))
+              onClick={() => {
+                onPresetChange(item.title);
+                dispatch(setPresetListVisible(false));
               }}
             >
               <Skeleton avatar title={false} loading={item.loading} active>
@@ -56,7 +58,7 @@ export function Preset(props: Props) {
         />
       </div>
     </>
-  ) : null
+  ) : null;
 }
 
 const styles = {
@@ -64,4 +66,4 @@ const styles = {
     // display: 'flex',
     // flexDirection: 'row',
   },
-}
+};
