@@ -1,4 +1,4 @@
-import { createSlice, PayloadAction } from '@reduxjs/toolkit'
+import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 
 interface SettingModule {
   visible: boolean;
@@ -6,18 +6,18 @@ interface SettingModule {
 
 export const initialState: SettingModule = {
   visible: false,
-}
+};
 
 export const settingSlice = createSlice({
   name: 'setting',
   initialState,
   reducers: {
     setVisible: (state, action: PayloadAction<boolean>) => {
-      const { payload } = action
-      state.visible = payload
+      const { payload } = action;
+      state.visible = payload;
     },
   },
-})
+});
 
-export const { setVisible } = settingSlice.actions
-export default settingSlice.reducer
+export const { setVisible } = settingSlice.actions;
+export default settingSlice.reducer;
