@@ -1,10 +1,12 @@
-import ReactDOM from "react-dom";
+import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import store from './app/store';
-import { App } from "./App";
+import { App } from './App';
 
-ReactDOM.render((
-  <Provider store={store}>
-    <App />
-  </Provider>
-), document.getElementById("root"));
+ReactDOM.render(
+  (
+    <Provider store={store}>
+      <App />
+    </Provider>
+  ), document.getElementById('root'),
+);
